Updated the function calls
The change was done because of no output. The UART2_Init was called then HAL_UART_Transmit and then the clock source was configured for PLL and then the UART2_Init was called then HAL_UART_Transmit again. The Systick timer was configured in the SystemClockConfig so it was not done explicitly in the main.c again